+++ src/llm-providers/_common/utils/count-total-usage/countTotalUsage.ts
@@ -1,16 +1,13 @@
 import type { Promisable } from 'type-fest';
 import type { AvailableModel } from '../../../../execution/AvailableModel';
 import type { LlmExecutionTools } from '../../../../execution/LlmExecutionTools';
-import type { ChatPromptResult } from '../../../../execution/PromptResult';
-import type { CompletionPromptResult } from '../../../../execution/PromptResult';
-import type { EmbeddingPromptResult } from '../../../../execution/PromptResult';
+import type { ChatPromptResult, CompletionPromptResult, EmbeddingPromptResult } from '../../../../execution/PromptResult';
 import type { PromptResultUsage } from '../../../../execution/PromptResultUsage';
 import { addUsage } from '../../../../execution/utils/addUsage';
 import { ZERO_USAGE } from '../../../../execution/utils/usage-constants';
-import type { ChatPrompt } from '../../../../types/Prompt';
-import type { CompletionPrompt } from '../../../../types/Prompt';
-import type { EmbeddingPrompt } from '../../../../types/Prompt';
+import type { ChatPrompt, CompletionPrompt, EmbeddingPrompt } from '../../../../types/Prompt';
 import type { LlmExecutionToolsWithTotalUsage } from './LlmExecutionToolsWithTotalUsage';
+import { Subject, type Observable } from 'rxjs';
 
 /**
  * Intercepts LLM tools and counts total usage of the tools
@@ -19,8 +16,9 @@ import type { LlmExecutionToolsWithTotalUsage } from './LlmExecutionToolsWithTot
  * @returns LLM tools with same functionality with added total cost counting
  * @public exported from `@promptbook/core`
  */
-export function countTotalUsage(llmTools: LlmExecutionTools): LlmExecutionToolsWithTotalUsage {
+export function countUsage(llmTools: LlmExecutionTools): LlmExecutionToolsWithTotalUsage {
     let totalUsage: PromptResultUsage = ZERO_USAGE;
+    const spending = new Subject();
 
     const proxyTools: LlmExecutionToolsWithTotalUsage = {
         get title() {
@@ -41,6 +39,10 @@ export function countTotalUsage(llmTools: LlmExecutionTools): LlmExecutionToolsW
             return /* not await */ llmTools.listModels();
         },
 
+        spending(): Observable<PromptResultUsage>  {
+          return spending;
+      },
+
         getTotalUsage() {
             // <- Note: [🥫] Not using getter `get totalUsage` but `getTotalUsage` to allow this object to be proxied
             return totalUsage;
@@ -52,6 +54,7 @@ export function countTotalUsage(llmTools: LlmExecutionTools): LlmExecutionToolsW
             // console.info('[🚕] callChatModel through countTotalUsage');
             const promptResult = await llmTools.callChatModel!(prompt);
             totalUsage = addUsage(totalUsage, promptResult.usage);
+            spending.next(promptResult.usage);
             return promptResult;
         };
     }
@@ -61,6 +64,7 @@ export function countTotalUsage(llmTools: LlmExecutionTools): LlmExecutionToolsW
             // console.info('[🚕] callCompletionModel through countTotalUsage');
             const promptResult = await llmTools.callCompletionModel!(prompt);
             totalUsage = addUsage(totalUsage, promptResult.usage);
+            spending.next(promptResult.usage);
             return promptResult;
         };
     }
@@ -70,6 +74,7 @@ export function countTotalUsage(llmTools: LlmExecutionTools): LlmExecutionToolsW
             // console.info('[🚕] callEmbeddingModel through countTotalUsage');
             const promptResult = await llmTools.callEmbeddingModel!(prompt);
             totalUsage = addUsage(totalUsage, promptResult.usage);
+            spending.next(promptResult.usage);
             return promptResult;
         };
     }