+++ src/types/typeAliases.ts
@@ -516,14 +516,14 @@ export type string_uuid = string & {
  *
  * @@@
  */
-export type string_app_id = id;
+export type string_app_id = id | 'app';
 
 /**
  * End user identifier
  *
  * @@@
  */
-export type string_user_id = id;
+export type string_user_id = id | string_email;
 
 /**
  * Semantic helper