+++ CHANGELOG.md
@@ -724,7 +724,7 @@ Use `.book` as default extension for books
 -   Use [VCCode extension `Promptbook .book language support`](https://marketplace.visualstudio.com/items?itemName=promptbook.book-extension)
 -   Fix: Version published to Docker Hub does not lag behind the NPM version
 
-### `0.88.0` _(2025-03-16)_
+### `0.88.0` _(2025-03-19)_
 
 Scripting and execution
 
@@ -738,6 +738,13 @@ Scripting and execution
 -   Add a unique ID to the error, this error needs to be serialised and deserialised.
 
 ## In pre-release
+
+### `0.89.0` _(2025-03-)_
+
+User and spending of credits
+
+-   Add User token to application mode on remote server
+-   Rename `countTotalUsage` -> `countUsage` and add `spending()`
 
 ## Drafts